remove & extract duplicates
