+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetFragment.scala
@@ -1,0 +1,45 @@
+package x7c1.linen.modern.init.settings.preset
+
+import android.app.Activity
+import android.support.v4.app.Fragment
+import x7c1.linen.glue.activity.ActivityControl
+import x7c1.linen.glue.res.layout.SettingPresetChannelRow
+import x7c1.linen.glue.service.ServiceControl
+import x7c1.linen.modern.accessor.LinenOpenHelper
+import x7c1.linen.modern.accessor.setting.PresetChannelsAccessor
+import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
+import x7c1.wheat.macros.logger.Log
+
+trait PresetFragment { self: Fragment =>
+
+  def reload(channelId: Long): Unit
+
+  def activity: Activity with ActivityControl with ServiceControl =
+    getActivity.asInstanceOf[Activity with ActivityControl with ServiceControl]
+
+  protected def args: PresetFragmentArguments
+
+  protected lazy val helper = new LinenOpenHelper(getContext)
+
+  protected lazy val presetsAccessor = {
+    PresetChannelsAccessor.create(args.accountId, helper) match {
+      case Right(accessor) => Some(accessor)
+      case Left(error) =>
+        Log error error.toString
+        None
+    }
+  }
+  protected def toAdapter
+    (location: PresetEventLocation)(accessor: PresetChannelsAccessor) = {
+
+    val factory = new PresetsChannelsAdapterFactory(
+      activity, args.rowFactory, location, helper, args.accountId
+    )
+    factory.createAdapter(accessor)
+  }
+}
+
+trait PresetFragmentArguments{
+  def accountId: Long
+  def rowFactory: ViewHolderProviderFactory[SettingPresetChannelRow]
+}
