+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetsAllFragment.scala
@@ -1,16 +1,11 @@
 package x7c1.linen.modern.init.settings.preset
 
-import android.app.Activity
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.support.v7.widget.LinearLayoutManager
 import android.view.{LayoutInflater, View, ViewGroup}
-import x7c1.linen.glue.activity.ActivityControl
 import x7c1.linen.glue.res.layout.{SettingPresetChannelRow, SettingPresetTabAll}
-import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.modern.accessor.database.ChannelSubscriber
-import x7c1.linen.modern.accessor.setting.PresetChannelsAccessor
 import x7c1.linen.modern.accessor.{AccountIdentifiable, LinenOpenHelper}
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
 import x7c1.wheat.macros.fragment.TypedFragment
@@ -22,59 +17,11 @@ class ArgumentsForAll(
   val accountId: Long,
   val tabFactory: ViewHolderProviderFactory[SettingPresetTabAll],
   val rowFactory: ViewHolderProviderFactory[SettingPresetChannelRow]
-)
+) extends PresetFragmentArguments
 
-trait ReloadableFragment { self: Fragment =>
-
-  def reload(channelId: Long): Unit
-
-  def activity: Activity with ActivityControl with ServiceControl =
-    getActivity.asInstanceOf[Activity with ActivityControl with ServiceControl]
-
-  protected def accountId: Long
-
-  protected lazy val helper = new LinenOpenHelper(getContext)
-
-  protected lazy val presetsAccessor = {
-    PresetChannelsAccessor.create(accountId, helper) match {
-      case Right(accessor) => Some(accessor)
-      case Left(error) =>
-        Log error error.toString
-        None
-    }
-  }
-}
-
-class PresetsChannelsAdapterFactory(
-  activity: Activity with ActivityControl with ServiceControl,
-  factory: ViewHolderProviderFactory[SettingPresetChannelRow],
-  location: PresetEventLocation,
-  helper: LinenOpenHelper, accountId: Long){
-
-  def createAdapter(accessor: PresetChannelsAccessor): PresetsChannelsAdapter = {
-    new PresetsChannelsAdapter(
-      listener = new SubscriptionChangedUpdater(accountId, activity, helper),
-      onSourceSelected = new OnSourcesSelected(activity).transitToSources,
-      onMenuSelected = new OnMenuForSelected(activity, accountId),
-      accessor = accessor,
-      provider = factory create activity,
-      location = location
-    )
-  }
-}
-
-class PresetsAllFragment extends TypedFragment[ArgumentsForAll] with ReloadableFragment {
-  private lazy val args = getTypedArguments
-
+class PresetsAllFragment extends TypedFragment[ArgumentsForAll] with PresetFragment {
+  protected lazy val args = getTypedArguments
   private lazy val layout = args.tabFactory.createViewHolder(getView)
-
-  private def toAdapter(accessor: PresetChannelsAccessor) = {
-    val factory = new PresetsChannelsAdapterFactory(
-      activity, args.rowFactory, PresetTabSelected, helper, args.accountId
-    )
-    factory.createAdapter(accessor)
-  }
-  override protected def accountId: Long = args.accountId
 
   override def reload(channelId: Long) = {
     Log info s"[start] $channelId"
@@ -86,7 +33,7 @@ class PresetsAllFragment extends TypedFragment[ArgumentsForAll] with ReloadableF
 
     Log info s"[start]"
     val tab = args.tabFactory.create(inflater).inflateOn(container)
-    presetsAccessor map toAdapter foreach { adapter =>
+    presetsAccessor map toAdapter(PresetTabAll) foreach { adapter =>
       tab.channelList setLayoutManager new LinearLayoutManager(getContext)
       tab.channelList setAdapter adapter
     }
