+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetsChannelsAdapter.scala
@@ -1,11 +1,15 @@
 package x7c1.linen.modern.init.settings.preset
 
+import android.app.Activity
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.{View, ViewGroup}
+import x7c1.linen.glue.activity.ActivityControl
 import x7c1.linen.glue.res.layout.SettingPresetChannelRow
+import x7c1.linen.glue.service.ServiceControl
+import x7c1.linen.modern.accessor.LinenOpenHelper
 import x7c1.linen.modern.accessor.setting.{SettingPresetChannel, PresetChannelsAccessor}
 import x7c1.linen.modern.display.settings.ChannelSourcesSelected
-import x7c1.wheat.ancient.resource.ViewHolderProvider
+import x7c1.wheat.ancient.resource.{ViewHolderProviderFactory, ViewHolderProvider}
 import x7c1.wheat.modern.decorator.Imports._
 
 class PresetsChannelsAdapter(
@@ -59,4 +63,22 @@ object MenuSelected {
 
 trait OnMenuSelectedListener {
   def onMenuSelected(e: MenuSelected): Unit
+}
+
+class PresetsChannelsAdapterFactory(
+  activity: Activity with ActivityControl with ServiceControl,
+  factory: ViewHolderProviderFactory[SettingPresetChannelRow],
+  location: PresetEventLocation,
+  helper: LinenOpenHelper, accountId: Long){
+
+  def createAdapter(accessor: PresetChannelsAccessor): PresetsChannelsAdapter = {
+    new PresetsChannelsAdapter(
+      listener = new SubscriptionChangedUpdater(accountId, activity, helper),
+      onSourceSelected = new OnSourcesSelected(activity).transitToSources,
+      onMenuSelected = new OnMenuForSelected(activity, accountId),
+      accessor = accessor,
+      provider = factory create activity,
+      location = location
+    )
+  }
 }
