+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetsSelectedFragment.scala
@@ -8,32 +8,22 @@ import android.view.{LayoutInflater, Menu, MenuItem, View, ViewGroup}
 import x7c1.linen.glue.res.layout.{SettingPresetChannelRow, SettingPresetTabSelected}
 import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.glue.service.ServiceLabel.Updater
-import x7c1.linen.modern.accessor.setting.PresetChannelsAccessor
 import x7c1.linen.modern.init.updater.UpdaterMethods
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
 import x7c1.wheat.macros.fragment.TypedFragment
 import x7c1.wheat.macros.intent.ServiceCaller
 import x7c1.wheat.macros.logger.Log
 
-
 class ArgumentsForSelected(
   val accountId: Long,
   val tabFactory: ViewHolderProviderFactory[SettingPresetTabSelected],
   val rowFactory: ViewHolderProviderFactory[SettingPresetChannelRow]
-)
-class PresetsSelectedFragment extends TypedFragment[ArgumentsForSelected] with ReloadableFragment {
-  private lazy val args = getTypedArguments
+) extends PresetFragmentArguments
 
+class PresetsSelectedFragment extends TypedFragment[ArgumentsForSelected] with PresetFragment {
+  protected lazy val args = getTypedArguments
   private lazy val layout = args.tabFactory.createViewHolder(getView)
 
-  override protected def accountId = args.accountId
-
-  private def toAdapter(accessor: PresetChannelsAccessor) = {
-    val factory = new PresetsChannelsAdapterFactory(
-      activity, args.rowFactory, PresetTabSelected, helper, args.accountId
-    )
-    factory.createAdapter(accessor)
-  }
   override def reload(channelId: Long) = {
     Log info s"[start] $channelId"
     presetsAccessor foreach (_.reload())
@@ -44,7 +34,7 @@ class PresetsSelectedFragment extends TypedFragment[ArgumentsForSelected] with R
 
     Log info s"[start]"
     val tab = args.tabFactory.create(inflater) inflateOn container
-    presetsAccessor map toAdapter foreach { adapter =>
+    presetsAccessor map toAdapter(PresetTabSelected) foreach { adapter =>
       tab.channelList setLayoutManager new LinearLayoutManager(getContext)
       tab.channelList setAdapter adapter
     }
